+++ src/Livewire/Sitemap.php
@@ -7,6 +7,7 @@ namespace Arkhe\Main\Livewire;
 use Arkhe\Main\Contracts\SiteSeoRepositoryInterface;
 use Arkhe\Main\Jobs\GenerateSitemap;
 use Arkhe\Main\Services\SitemapService;
+use Flux\Flux;
 use Illuminate\Contracts\View\View;
 use Livewire\Component;
 
@@ -28,7 +29,7 @@ class Sitemap extends Component
 
         GenerateSitemap::dispatch();
 
-        session()->flash('arkhe.sitemap.dispatched', true);
+        Flux::toast(variant: 'success', text: __('arkhe::arkhe.sitemap.dispatched'));
     }
 
     public function render(